refactor(pdp): decouple PDP into smaller components
